Reorganize exports for fewer and more succint imports in real use
This one's aimed at moving some types/files around so that a program
using the package has a more succinct set of River imports. Types that
are meant to be public-facing largely end up in `riverqueue/__init__.py`,
and more internal types get moved elsewhere.

So what used to be this:

    from riverqueue.client import Client
    from riverqueue.models import InsertOpts, UniqueOpts
    from riverqueue.drivers.sqlalchemy.sqlalchemy_driver import SqlAlchemyDriver

    client = Client(SqlAlchemyDriver(engine))

    result = client.insert(
        MyJobArgs(),
        insert_opts=InsertOpts(unique_opts=UniqueOpts(by_period=900)),
    )

Becomes:

    from riverqueue import Client, InsertOpts, UniqueOpts
    from riverqueue.driver import riversqlalchemy

    client = Client(riversqlalchemy.Driver(engine))

    result = client.insert(
        MyJobArgs(),
        insert_opts=InsertOpts(unique_opts=UniqueOpts(by_period=900)),
    )

(It's a subtle change to a degree, but see the import lines in
particular.)

There's a few other changes to clean up naming somewhat:

* `sqlalchemy_driver.SqlAlchemyDriver` -> `riversqlalchemy.Driver`
* `riverqueue.drivers` -> `riverqueue.driver` (plural to singular) to
  better match up other River projects.
* `riverqueue.models` -> `riverqueue.model` (plural to singular) to
  match the change for drivers.
* `Driver` -> `DriverProtocol` so there's no ambiguity problems where
  both it and a concrete driver implementation are imported together.
